Example for MAE 223 on 13-2-2012
A body A rotating in space, with an 'L' shaped arm rotating about point O along
the A_1 axis, the angle of rotation measured by theta. There is a point P
moving on the "forearm" part of the arm.

A and O have predefined velocities. Velocity and acceleration of P is found in
N using 1 and 2 point theorems.