Added rolling disc example.
This approach uses with minimal set of generalized coordinates and speeds and
forms the constraint forces.  Comparing these results to the results of
KanesMethod won't indicate whether the handling of dependent speeds or
configuration constraints are being handled correctly because the formulation
is such that there is no constraint equation to solve.