Added nonminimal rolling disc example.
This example derives the equations of motion and constraint forces of the
rolling disc when using a non-minimal choice of generalized coordinates and a
nonminimal choice of generalized speeds.  The nonholonomic generalized inertia
forces are mathematically equal to those produced by rollingdisc.py, but there
are a few trig identities you have to implement inorder to make the symbolic
comparison assert as true.  Using trigsimp even with deep=True and
recursive=True didn't fix this, so perhaps the best thing to do would just
evaluate the equations at a few lean different lean angles q[1] which have
exact known solutions like pi/3, pi/4 or something, and then that way we can
still get a symbolic comparison on everything else in the expressions.