Use different subprocess command for Cython.
Previously, we used a combination of `Popen()` and `wait()` to compile
a Cython file but suppress stdout. On my machine, code continued to
execute despite the call to `wait()`. If I use `call()` instead, the
code waits until the compiling is finished before it proceeds. This is
important for ensuring the tests are executed properly.