Modified .travis.yml to install pip<6, setuptools<8 to avoid PEP 440 errors.
SymPy uses a development version idenitifier that looks like `0.7.6-git`. This
does not comply with PEP 440 and both pip>6 and setuptools>8 enforce PEP 440,
causing any version with a trailing `-git` to be treated as the oldest version.
Since we test against SymPy master, this is a problem because `python setup.py
install` will not accept SymPy version `0.7.6-git` as newer than `0.7.4.1` and
the function `utils.sympy_version_newer_than` fails because it relies on
`pkg_resources.parse_version`. So the quick fix is to install earlier versions
of pip and setuptools on Travis and in your development environment, i.e. any
time you are using development versions of SymPy to run PyDy.

- Some .travis.yml reshuffling and more version printing.
- Added a nose upgrade command.
- Added a test for utils.sympy_version_newer_than()