Create Cycling leg
The file shows a DOF system representing a leg that makes a cycling motion. The dynamics of the rotating pedal is currently dependent on the dynamics of the leg. I am still searching for a solution to make the pedal rotate with a constant velocity; making it a 1 DOF system.