Update matrix_generator.py
The remove function from the set class raises an error when the argument is not in the set. I have changed it to the discard function, which does the same thing but does not raise the error. This is useful if you create custom matrices that do not have the symbol t.