Added Whipple-Carvallo documentation example.
This is based off the whipple.py. The simulation works and is stable. I
get divide by zeros and nan's if the initial conditions are zero. The
small epsilon I added in place of the zero initial conditions worked for
1e-12 but not 1e-10 and 1e-14. Not sure why either of these issues
occur. But at least this is a basic working example and a starting point
for investigating more.