Added a symbolic linear solver option to the CythonODEFunctionGenerator.
This feels a little hacky as is due to the code being a little confusing
with the patched CMatrixGenerator and returning A,x instead of A,b from
Ax=b. But it seems to work and should allow some larger expressions and
matrices to actually compile in a more reasonable time if you want a
symbolically solved system.

Needs another test and some cleanup still.

Fixes #450.